feat: Extract NavigationList and NavigationListSection and expose them
