+++ docs/styleguide.config.js
@@ -113,7 +113,8 @@ module.exports = {
         '../react/AppLinker/index.jsx',
         '../react/Breadcrumbs/index.jsx',
         '../react/Menu/index.jsx',
-        '../react/Tabs/index.jsx'
+        '../react/Tabs/index.jsx',
+        '../react/NavigationList/index.jsx'
       ]
     },
     {
